backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-sqlalchemy>=3.1.0
bcrypt>=4.1.0
groq>=0.4.0
pandas>=2.0.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
werkzeug>=3.0.0
torch>=2.0.0